requirements.txt
glfw 
PyOpenGL
